temporalized/requirements.txt
temporalio>=1.6.0
boto3>=1.34.0
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
tqdm>=4.64.0
